requirements-test.txt
pytest
pytest-datadir
pytest-cov
codecov
python-telegram-bot
environs
requests
pronouncing
pybadges
nltk
google_spreadsheet
google-auth-oauthlib
panadas
azure-cognitiveservices-speech